# openapi.storefront.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Commerce Platform Storefront API",
    "description": "Theme-facing page models and cart (spec §8.2). Reached through the edge only.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/storefront/v1/customer/addresses": {
      "get": {
        "tags": [
          "storefront"
        ],
        "operationId": "list_addresses",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Customer-Session",
            "in": "header",
            "description": "Session token from the checkout origin's `__Host-sid` cookie (set by the edge).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressList"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "storefront"
        ],
        "operationId": "add_address",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Customer-Session",
            "in": "header",
            "description": "Session token from the checkout origin's `__Host-sid` cookie (set by the edge).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Address"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/customer/addresses/{id}": {
      "put": {
        "tags": [
          "storefront"
        ],
        "operationId": "update_address",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Customer-Session",
            "in": "header",
            "description": "Session token from the checkout origin's `__Host-sid` cookie (set by the edge).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Address"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "storefront"
        ],
        "operationId": "delete_address",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Customer-Session",
            "in": "header",
            "description": "Session token from the checkout origin's `__Host-sid` cookie (set by the edge).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {},
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/customer/login": {
      "post": {
        "tags": [
          "storefront"
        ],
        "operationId": "login",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Checkout cart capability (`__Host-cart`): the cart is attached to the customer (A4).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Consent-Subject",
            "in": "header",
            "description": "Anonymous consent subject (consent cookie): its choices move to the customer (A20).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Client-Ip",
            "in": "header",
            "description": "The client's IP (rate limits; stored only as a salted hash).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "headers": {
              "X-Session-Token": {
                "schema": {
                  "type": "string"
                },
                "description": "New session (for the edge only)"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResult"
                }
              }
            }
          },
          "401": {
            "description": "invalid_credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "too_many_attempts",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/customer/logout": {
      "post": {
        "tags": [
          "storefront"
        ],
        "operationId": "logout",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Customer-Session",
            "in": "header",
            "description": "Session token from the checkout origin's `__Host-sid` cookie (set by the edge).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "headers": {
              "X-Session-Clear": {
                "schema": {
                  "type": "string"
                },
                "description": "The edge deletes the cookie"
              }
            }
          }
        }
      }
    },
    "/storefront/v1/customer/magic-link": {
      "post": {
        "tags": [
          "storefront"
        ],
        "operationId": "request_magic_link",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Checkout cart capability (`__Host-cart`): the cart is attached to the customer (A4).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Consent-Subject",
            "in": "header",
            "description": "Anonymous consent subject (consent cookie): its choices move to the customer (A20).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Client-Ip",
            "in": "header",
            "description": "The client's IP (rate limits; stored only as a salted hash).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Sent if the address can receive mail (same answer for every address)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "too_many_attempts",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressInput": {
        "type": "object",
        "required": [
          "name",
          "street",
          "city",
          "postal_code",
          "country"
        ],
        "properties": {
          "city": {
            "type": "string"
          },
          "company": {
            "type": [
              "string",
              "null"
            ]
          },
          "country": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "postal_code": {
            "type": "string"
          },
          "street": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AddressList": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details body.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code, e.g. `not_found`."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`; `code` carries the specific error kind."
          }
        }
      },
...
    }
  },
...
}